Fix int4 dispatch + scale dtype inconsistency (P6b / P4b)
FloatTensor::dotQInteger CPU fallback now calls
nntr_gemm_qai8dxp_qsi4cxp_unpacked<float>() instead of the _packed
variant. The _packed variant requires a KleidiAI pre-packed buffer
(tile-interleaved with integrated fp32 scales) and silently produced
garbage on ARM / threw NYI on x86 when called with Int4QTensor's raw
canonical layout. The _unpacked variant takes the canonical shape
directly:

    rhs_native_mtx_qs4cx = packed nibbles (Int4QTensor::getData())
    rhs_scales_f32       = fp32 scale array (Int4QTensor::getScale())

and does its own runtime packing internally, so the Tensor dispatch
path is finally usable without a separate repack step. A follow-up
can add per-instance caching of a pre-packed buffer if throughput
profiling shows the runtime-pack overhead is significant.

The unpacked variant also has a working fallback path in
fallback_internal_fp16.cpp::__fallback_nntr_gemm_qai8dxp_qsi4cxp_unpacked
which implements reference matmul via ref_quant_qa8dx_f32 +
ref_matmul_f32_qa8dx_qs4cx, so x86 builds no longer throw NYI on this
code path.

P4b latent bug fixed in the same commit for coherence:
Int4QTensor::allocate() has always reserved sizeof(float)*scale_size()
bytes for the scale section (line 144), but getMemoryBytes() reported
scale_size*sizeof(uint16_t). save() and read() both use
getMemoryBytes() for the byte count, which meant save() was only
writing the low 2 bytes of each fp32 scale while the allocated buffer
held correct fp32 values. On reload, scales were garbage. This never
surfaced because no test exercised a real save+load round-trip for
QINT4 weights through the NeuralNetwork path, and the old FC dequantize
shortcut worked entirely in memory without hitting the file. Switch
getMemoryBytes() to sizeof(float) so save/read match the actual
allocation, and document the fp32 canonical on the class header.
This also matches the KleidiAI qai8dxp_qsi4cxp_unpacked interface
(which takes fp32 rhs_scales) and the getScale<float>() access
pattern already used by char_tensor.cpp and quantizer.cpp.

Added Q4_0 regression smoke tests to
unittest_nntrainer_safetensors_int4.cpp so future refactors to the
QINT4 branches cannot silently break Lane B block-quantized weights:

  SafetensorsQ4_0.BuildFcWithQ4_0Weight_smoke
    Verifies that an FC layer with weight_dtype=Q4_0 (K=16, N=32)
    compiles and initializes cleanly. Guards against P5-style
    changes that might accidentally null out the default branch in
    FullyConnectedLayer::finalize or remove Q4_0 from the Tensor
    factory.

  SafetensorsQ4_0.SaveQ4_0Fc_header_contains_q4_0_quant
    Saves the NN and spot-checks the safetensors JSON header for
    schema_version 2, dtype="Q4_0", quant.encoding="q4_0",
    bitwidth=4, group_size=32. Guards the P2 Lane B dtype table
    (commit 807f18d).

  SafetensorsQ4_0.SaveThenLoadQ4_0Fc_roundtrip
    Full save+load round-trip through NeuralNetwork. The P3a
    validator must accept "Q4_0" as a quantized dtype (c_is_quant
    includes Q4_0 since commit 807f18d) and must not flag the quant
    object as a mismatch against Q4_0_Tensor.

Q4_0 is on Lane B (dispatched via FloatTensor::dotQnK ->
o->gemm_q4_0_fp32 or gemm_q4_0_accel_fp32) and is not touched by
the dotQInteger changes above.

+++ nntrainer/tensor/int4_tensor.cpp
@@ -550,8 +550,19 @@ void Int4QTensor::print(std::ostream &out) const {
 }
 
 size_t Int4QTensor::getMemoryBytes() const {
+  // Scales are stored as fp32 (sizeof(float)) per the canonical layout
+  // documented on the class header. allocate() already reserves
+  // `sizeof(float) * scale_size()` bytes for the scale section, and the
+  // KleidiAI qai8dxp_qsi4cxp_unpacked kernel consumes fp32 scales. Before
+  // P6b this function reported `sizeof(uint16_t) * scale_size()` which
+  // caused save() and read() to transfer only the low 2 bytes of each
+  // fp32 scale (= garbage), while the in-memory buffer actually held
+  // correct fp32 values. That silent data corruption on round-trip
+  // never surfaced because the FC layer's old dequantize path worked
+  // entirely in memory and the test coverage for saved-then-reloaded
+  // QINT4 weights was effectively zero. Align with the allocator here.
   return ((size() + 1) / 2) * dim.getDataTypeSize() +
-         scale_size() * sizeof(uint16_t);
+         scale_size() * sizeof(float);
 }
 
 size_t Int4QTensor::scale_size() const {
